Fix argument parsing without script bug
The ArgumentParser would check if the parse argument was an option, if it
was it would then use the default script convention and continue parsing
options. However, it would never actually parse the first option, so for
example:

    cake.exe -option1=value1 -option2=value2

Would result in `option2` being successfully parsed, while `option1` is
silently dropped.

I believe this went unnoticed because all the unit tests and the default
bootstrapper script explicitly pass in the build script.