Add new globber.
The old globber was slow, complicated and difficult
to debug so I decided to rewrite it from scratch.

From my benchmarks there's a ~16X increase in speed,
and I also found bugs in the old implementation that
I've added tests for. There should be no difference
in behavior according to the tests we had in place.