Added using-alias-directive support.
Seems to work correctly with Roslyn, but on Mono there seem to
be a bug in the Evaluator that throws a NullReferenceException.
Since the exception is so vague, I added a better error message
for the user.

I've filed a bug report with Xamarin for this that can
be found at https://bugzilla.xamarin.com/show_bug.cgi?id=32894#c0

Resolves #328