Restored accidental sematic change with globber predicates.
Before the Globber rewrite, the predicate was used as a way of
telling the Globber what directories to recurse into when using
double star (`**`) or directory wildcard ('*'). When rewriting
the globber, the predicate was used to determine filtering of
directories and files in the results. This has now been changed.
This will (sadly) lead to a breaking change for people who rely
on filtering the results, but this could be easily corrected by
simply using a `Where` clause on the resulting collection.