Added verbose message when glob pattern did not match anything.
Sometimes when providing a glob pattern to certain aliases such
as NUnit or XUnit, a mistake in the pattern will result in
zero files being matched. This is kind of annoying since the
call to the alias will silently pass. This might be by design
but we should at least write a verbose message saying that no
files or directories were matched.

Closes #408