Improve argument parser to use unix-arguments
The original version used a non-conventional single-dash argument as a way
of passing arguments. This has been changed to double-dash arguments for
long options and single-dash argument for single character options.

To prevent this commit from becoming a breaking change the old behaviour
is also supported. It might be removed in the future in favor of the new
style, however until documentation and features have been updated this
should have no effect for everyone.

In the new approach we have also introduced a new way to create additional
options in `CakeOptions`. The command line key name is now decided by an
attribute. In the future we might wanna expend that with documentation
attribute so that we can generate the help on the fly. (Also it would be
really cool if we could let cake parse the build file for additional
options).

All the newly introduced code has been written using TDD behaviour and
exception for two error messages (so that it would support the new
behaviour) none of the original UnitTests have been changed.