Enabled tests for POSIX based systems.
I needed to make some changes to get this to work such
as updating xUnit to version 2.1.0 and enabling the
`NoAppDomain` flag for the xUnit runner. This turned out
to reduce the test time with 75%, but also lead to that we
had to change some test code to make it run properly.

I've also created two new xUnit attributes called
`WindowsFact` and `WindowsTheory` that will be skipped if
the tests are not running on a Windows machine. Not ideal
but this should only be used for platform specific tests.
We should probably look into removing these tests or rewrite
the underlying mechanism that's under test so we don't rely
on platform specific tests in the future, but for now it
will have to do.

Closes #575