Make #tool and #addin package manager agnostic.
This is a new way of referencing packages. This so we
can allow the usage of other package managers such as npm and
homebrew to retrieve packages in the future.

We still support the old #addin/#tool directive, but the new
syntax for it is more generic and easy to expand. We will
however need to update the documentation for this asap.

There are some breaking changes though. I've moved all the
NuGet stuff to it's own assembly (except INuGetToolResolver)
since those aren't really of interest for consumers of
Cake.Core. Keeping this isolated will potentially make it
easier when migrating to CoreCLR in the future as well.

nuget:?package=Cake.Foo
nuget:https://myget.org/f/Cake/?package=Cake.Foo&Version=1.2.3

Closes #616