Built in parameters now accept boolean values
This PR enabled the built in parameters to Cake.exe
to take the form as a boolean key/value pair instead
of only as a flag.

This means that `--debug=true` will be the same thing
as `--debug` and will give the user an opportunity to
explicitly tell Cake to NOT debug the script by
passing `--debug=false`. This is useful when writing
a custom bootstrapper for Cake.

Closes #917