Refactored ICakeEnvironment
This included obsoleting a lot of members and also
introducing two concepts, ICakePlatform and ICakeRuntime.

ICakePlatform is an abstraction around the actual platform
that Cake is running on. This will give you information about
things such the operating system and architecture.

ICakeRuntime is an abstraction around the runtime that Cake
is running in such as the .NET target platform.

Closes #502