Added modularity changes to the core. Refactored LineProcessor -> Split() to its own extension both one for LineProcessorExtensions and StringExtensions. All processors that used the Split has moved its calls to the LineProcessor extension. Added a new method for IScriptProcessor called InstallPackage this takes care of installations of PackageReferences. Made HashSetExtension public and added documentation to it. Fixed the existing test cases for the new IProcessorExtensionFinder.
Checking for a processor extension in ScriptAnalyzer.cs line 135. this will hook any registered processo extension and process it if it can.

ProcessorExtension takes care of instructing a custom processor to get executed if the method CanProcessDIrective returns true.

ScriptRunnerExtension takes care of the actuall instruction on how to perform the installation.