GH-1166: refactor the MethodAliasGenerator to support optional parameters in alias methods
Refactor the parameter concerns in the method alias tokenizer to support
optional params, extract logic to ParameterEmitter class, add supporting tests.

Optional params have to be compile-time constants, so this keeps the surface area reasonable.
Nullable<> parameters can be optional, but the default value, if not null,
has to be cast as the inner type to be correct.

Add supporting tests for the various use cases.