default-paths: Use relative paths
The default path that is used for addin installation is being constructed
by using the application root. This is not the propagated default behavior
on the website. Let's use the relative path `./tools/addins` as default.

Also, those default paths should respect the current configured tools
path. Say you set the tools path to `./foo`, then the default modules and
addins paths would be `./foo/Addins` and `./foo/Modules`.

This fixes (GH-1344)