xmlpeek: Add switch to suppress warnings
The `XmlPeekSettings` learned to remember a setting if warnings should be
suppressed or not. That setting will be evaluated before calling
`context.Log.Warning()` if the `XmlPeek` alias can not find the provided
xpath.

With this new setting in place, users can now _opt out_ of displaying the
warning if the xpath was not found.

This fixes (GH-1438)