Added a more reliable method to detect Mac OSX
Previously on mono, the PlatformID enum was being used to detect OSX.  As of mono 2.2+ the value on Mac is always returned as 4, which makes it impossible to use to distinguish from any other unix based system.

The fix uses a method that’s tried and tested in MonoDevelop (Xamarin Studio & Visual Studio for Mac) calling out to `uname` and checking if the value is Darwin.