Fixed .NET Core parsing error when spaces in path
The donet.exe executable doesn't double quote the executable in the GetCommandLineArgs() method. When there are spaces is the path where cake.exe resides the prior version wouldn't parse the command line correctly. Fixed the code and pulled out some of the more complicated code so that you can simple call GetCommandLineArgs(). Also, added an optional parameter if the executable is wanted.