GH-2393: Second pass, attempt to more completely support codegen
As a follow-on to the initial commit, this attempts to inspect
those parameters decorated with attributes that utilize a more
complex constructor signature, supporting typical constructor arguments
and named arguments.

This attempts to generally mimic the logic exposed in CustomAttributeTypedArgument,
but using the normal Cake code-gen reflection utils to emit compile-safe type names, etc.