(GH-2638) Add ability to pass bool and null values
When creating CustomAttributes for AssemblyInfo files, it is useful to
be able to generate attributes with boolean values, as well as attributes
that don't have a value.  This is a "simple" solution to allow this to
happen, a broader fix would be to use Roslyn to fully understand all
attributes in the file.