Better handling for GitVersion failure cases
GitVersion was failing for our builds this week, but the reason was a mystery until we were able to dig into the GitVersion log file on a failed build. What made this really difficult is the actual error was masked by the GitVersionRunner.

The error according to Cake was `Expecting state 'Element'.. Encountered 'Text'  with name '', namespace ''.` This is accurate because the `DataContactJsonSerializer` is failing to deserialise what is actually an error message from `gitversion.exe`. However, it is ultimately misleading since the actual error message is masked by this exception.

We changed our process to call GitVersion twice:

1. Run GitVersion with `OutputType = GitVersionOutput.BuildServer` so it logs to stdout, including any error messages
2. Run GitVersion with `OutputType = GitVersionOutput.Json` to get the deserialised version information

This PR proposes a similar approach, but within the GitVersionRunner itself so others fall into the pit of success. It appears the intent of the code was to run GitVersion twice, but the deserialization exception prevents the intended behaviour.

Hope this helps!