(GH-2752) Fix precedence of VS2019 MSTest.exe to be higher than VS2017.
This re-orders the "Year and Edition" folder path segments of VS2019 to appear
first in the list as the newest VS version, which allows any alternative tool path
resolution to resolve latest first.  This follows the patterns established where the
resolution has historically resolved VS2015 over VS2013, and so on.