(#3858, #3859) Made PathCollapser aware of different formats
especially different calculations of the root of a path
for windows, non-windows and UNC like paths.

To this end, I re-structured the unit test. There were no test
removed that previously existed, but only new tests added.

Also, instead of joining together the path with one specific
separator ('/'), I re-used the existing separator of the original
path. (This was the main-fix for #3859.)