(GH-4055) Add Spectre.Console based reporter
Via a new setting, --Settings_UseSpectreConsoleForConsoleOutput, it is
now possible to output the task headers, and task summary using
Spectre.Console.  This allows for the easy addition of a new column in
the task summary, which includes information about _why_ any given task
has been skipped.

The old CakeReportPrinter is still in play, but some implementation has
been moved from within the DefaultExecutionStrategy to the reporter
class.  Depending on the setting mentioned above, either the
Spectre.Console version of the reporter will be added to the IoC
container, or the old version will be in place.

* fixes #4055