(#4475) Always show task execution summary
Currently, Cake prints out a task execution summary on each successful
execution.  This summary includes which tasks were executed or skipped,
and the duration of the task.  However, then an execution of Cake
fails, the task summary is not printed, and it is necessary to scroll
through the Cake output, to find out what task failed.  You also lose
the information about which tasks, if any, were skipped.

Any changes to this workflow will have to be done carefully though, as
we don't want to alter how tasks are executed, in what order, or what
is done in the cases of ContinueOnError, etc.  There are a number of
unit tests that check for this execution, and these will need to
continue to pass.

This commit aims to address this problem by introducing a new
CakeReportException, which has a property called Report of type
CakeReport.  That way, when an exception needs to be thrown, the
current execution status report can be passed along with it, and from
there, the task execution summary can be printed.

The introduction of the CakeReportException has meant that a number of
unit tests have had to be updated, since there is an expectation that
an InvalidOperationException will be returned, however, I think that
this is "ok", since it hasn't meant a change to how the workflow of
task executions completes.